Update README.md
### Extended Description

Welcome to my dynamic portfolio website, an all-in-one platform designed to help you showcase your skills, projects, and achievements in a modern and efficient way. Whether you're a developer, designer, or any professional looking to make a strong online presence, this portfolio template has got you covered.

#### Key Features:

- **Flexibility**: This portfolio is built using the latest tech stack, including Next.js, TypeScript, Sanity.io, Tailwind CSS, and deployed on Vercel. It offers a high degree of flexibility for customization, ensuring your portfolio truly represents your unique style and personality.

- **Easy Content Management**: With the power of Sanity.io, you can effortlessly manage and update your content. Add new projects, update skills, and edit your bio, all without the need for coding skills. Your portfolio remains fresh and up-to-date.

- **Responsive Design**: Tailwind CSS ensures that your portfolio looks stunning on all devices, from desktop to mobile. You won't have to worry about compatibility issues – your portfolio will shine everywhere.

- **Fast and Performant**: Next.js takes care of performance optimization. Your website will load quickly and efficiently, providing an excellent user experience.

- **Contribution-Friendly**: We encourage contributions! If you're interested in improving this project or adding new features, please follow our [Contributing Guidelines](CONTRIBUTING.md). Together, we can make this portfolio template even better.

#### Getting Started:

If you'd like to use this portfolio for your own professional online presence, follow these simple steps:

1. **Clone this Repository**: Start by cloning this repository to your local machine using `git clone`.

2. **Install Dependencies**: Use `npm install` to install all the necessary dependencies.

3. **Configuration**: Set up your portfolio by configuring the database and environment variables as detailed in the [Wiki](https://github.com/ritesh-15/my-portfolio/wiki).

4. **Customize**: Personalize your portfolio by updating the content, adding your projects, and tweaking the styling to match your preferences.

5. **Deploy**: Once you're satisfied with your changes, deploy your portfolio on your preferred hosting platform, whether it's Vercel, Netlify, or any other.

#### License:

This project is open-source and licensed under the MIT License. You are free to use, modify, and distribute it as per the terms of the license. For more details, please refer to the [LICENSE](LICENSE) file.

#### Contact:

If you have any questions, suggestions, or need assistance, feel free to reach out. Your feedback is invaluable in making this portfolio template even better.

Thank you for choosing this portfolio template for your online presence. We hope it serves you well and helps you stand out in the digital world!